long ?

+++ fewshotbench_v2/methods/metaoptnet_b.py
@@ -227,6 +227,6 @@ def one_hot(indices, depth):
 
     encoded_indicies = torch.zeros(indices.size() + torch.Size([depth])).cuda()
     index = indices.view(indices.size()+torch.Size([1]))
-    encoded_indicies = encoded_indicies.scatter_(1,index,1)
+    encoded_indicies = encoded_indicies.scatter_(1,index.long(),1)
     
     return encoded_indicies
